+++ apps/graduate/src/shared/constants/index.ts
@@ -1,1 +1,4 @@
 export * from './professors';
+export * from './endpoint';
+export { default as KEYS } from './keys';
+export { default as ROUTE } from './route';
