+++ apps/graduate/src/shared/constants/keys.ts
@@ -1,4 +1,6 @@
-export const KEYS = {
+const KEYS = {
   ACCESS_TOKEN: 'accessToken',
   REFRESH_TOKEN: 'refreshToken',
 } as const;
+
+export default KEYS;
