+++ apps/graduate/src/shared/utils/index.ts
@@ -1,2 +1,5 @@
-export { default as queryClient } from './queryClient';
 export * from './token';
+export { default as parseError } from './error';
+export { default as logout } from './logout';
+export { default as queryClient } from './queryClient';
+export { default as router } from './router';
